requirements.txt
asgiref==3.3.1
attrs==20.3.0
cffi==1.14.5
cryptography==3.4.7
da-vinci==0.3.0
Django==3.1.7
django-3-jet==1.0.8
django-appconf==1.0.4
django-extensions==3.1.1
django-filter==2.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.12.0
djangorestframework==3.12.4
drf-spectacular==0.14.0
inflection==0.5.1
jsonschema==3.2.0
pilkit==2.0
Pillow
pycparser==2.20
PyJWT==2.0.1
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
